+++ demo/admin/src/dam/ImportFromPicsum.tsx
@@ -1,7 +1,8 @@
-import { Button, CancelButton, messages, SaveButton } from "@comet/admin";
+import { Button, CancelButton, Loading, SaveButton } from "@comet/admin";
 import { Reload } from "@comet/admin-icons";
 import { useCurrentDamFolder, useDamAcceptedMimeTypes, useDamFileUpload } from "@comet/cms-admin";
 import {
+    Box,
     // eslint-disable-next-line no-restricted-imports
     Dialog,
     DialogActions,
@@ -26,13 +27,14 @@ export const ImportFromPicsum = () => {
     });
 
     const handleOpenDialog = async () => {
+        setIsOpen(true);
         const image = await getRandomPicsumImage();
         setPicsumImage(image);
-        setIsOpen(true);
     };
 
     const handleCloseDialog = () => {
         setIsOpen(false);
+        setPicsumImage(undefined);
     };
 
     const handleSave = async () => {
@@ -43,7 +45,6 @@ export const ImportFromPicsum = () => {
                 folderId,
             },
         );
-        handleCloseDialog();
     };
 
     const handleShuffle = async () => {
@@ -60,16 +61,21 @@ export const ImportFromPicsum = () => {
                 <div>
                     <DialogTitle>Import from Picsum</DialogTitle>
                     <DialogContent>
-                        <ImagePreview src={picsumImage?.url} alt="image" />
+                        <Box sx={{ aspectRatio: "16/9", lineHeight: 0 }}>
+                            {picsumImage ? <ImagePreview src={picsumImage?.url} alt="image" /> : <Loading behavior="fillParent" />}
+                        </Box>
                     </DialogContent>
                     <DialogActions>
                         <CancelButton onClick={handleCloseDialog} />
                         <Button variant="textDark" startIcon={<Reload />} onClick={handleShuffle}>
                             <FormattedMessage id="pages.dam.importFromPicsum.dialog.shuffle" defaultMessage="Shuffle" />
                         </Button>
-                        <SaveButton onClick={handleSave}>
-                            <FormattedMessage {...messages.save} />
-                        </SaveButton>
+                        <SaveButton
+                            onClick={async () => {
+                                await handleSave();
+                                handleCloseDialog();
+                            }}
+                        />
                     </DialogActions>
                 </div>
             </Dialog>