+++ demo/admin/src/products/future/ProductPriceForm.cometGen.ts
@@ -1,10 +1,10 @@
-import { type FormConfig } from "@comet/admin-generator";
+import { defineConfig } from "@comet/admin-generator";
 import { type GQLProduct } from "@src/graphql.generated";
 
-export const ProductPriceForm: FormConfig<GQLProduct> = {
+export default defineConfig<GQLProduct>({
     type: "form",
     gqlType: "Product",
     mode: "edit",
     fragmentName: "ProductPriceFormDetails", // configurable as it must be unique across project
-    fields: [{ type: "number", name: "price", helperText: "Enter price in this format: 123,45", startAdornment: "€" }],
-};
+    fields: [{ type: "number", name: "price", helperText: "Enter price in this format: 123.45", startAdornment: "€", decimals: 2 }],
+});