+++ packages/admin/cms-admin/src/dam/DataGrid/duplicatedFilenames/ManuallyHandleDuplicatedFilenamesDialog.tsx
@@ -1,8 +1,7 @@
-import { Alert } from "@comet/admin";
+import { Alert, Button } from "@comet/admin";
 import { Duplicate, Forward, Save } from "@comet/admin-icons";
 import {
     Box,
-    Button,
     // eslint-disable-next-line no-restricted-imports
     Dialog,
     DialogActions,
@@ -108,7 +107,7 @@ export const ManuallyHandleDuplicatedFilenamesDialog: VoidFunctionComponent<Dupl
             </DialogContent>
             <DialogActions>
                 <Box display="flex" justifyContent="space-between" width="100%">
-                    <Button variant="text" color="secondary" onClick={onSkip} startIcon={<Forward />}>
+                    <Button variant="textDark" onClick={onSkip} startIcon={<Forward />}>
                         <FormattedMessage
                             id="comet.dam.duplicateFilenameDialog.action.skip"
                             defaultMessage="Skip {count, plural, one {file} other {files}}"
@@ -127,7 +126,7 @@ export const ManuallyHandleDuplicatedFilenamesDialog: VoidFunctionComponent<Dupl
                                 }}
                             />
                         </Button>
-                        <Button variant="contained" color="primary" onClick={onUpload} startIcon={<Save />}>
+                        <Button onClick={onUpload} startIcon={<Save />}>
                             <FormattedMessage
                                 id="comet.dam.duplicateFilenameDialog.action.uploadAsCopy"
                                 defaultMessage="Upload as {count, plural, one {copy} other {copies}}"