+++ packages/admin/cms-admin/src/pages/SaveConflictDialog.tsx
@@ -1,8 +1,7 @@
-import { Alert, messages } from "@comet/admin";
+import { Alert, Button, messages } from "@comet/admin";
 import { Clear, Delete, OpenNewTab } from "@comet/admin-icons";
 import {
     Box,
-    Button,
     // eslint-disable-next-line no-restricted-imports
     Dialog,
     DialogActions,
@@ -96,7 +95,7 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                 </Stack>
             </DialogContent>
             <DialogActions>
-                <Button onClick={onClosePressed} startIcon={<Clear />} color="info">
+                <Button variant="textDark" onClick={onClosePressed} startIcon={<Clear />}>
                     <FormattedMessage {...messages.close} />
                 </Button>
                 <DialogActionsSpacer />
@@ -106,8 +105,6 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                         onClosePressed();
                         onDiscardChangesPressed();
                     }}
-                    variant="contained"
-                    color="info"
                 >
                     <FormattedMessage id="comet.saveConflictDialog.actionButtons.discardChanges" defaultMessage="Discard your changes" />
                 </DiscardButton>
@@ -117,8 +114,6 @@ function SaveConflictDialog({ open, onClosePressed, onDiscardChangesPressed }: S
                         onClosePressed();
                         window.open(window.location.href, "_blank");
                     }}
-                    variant="contained"
-                    color="primary"
                 >
                     <FormattedMessage
                         id="comet.saveConflictDialog.actionButtons.openCurrentVersionInNewTab"