+++ packages/admin/cms-admin/src/redirects/RedirectsGrid.tsx
@@ -1,5 +1,6 @@
 import { useQuery } from "@apollo/client";
 import {
+    Button,
     DataGridToolbar,
     FillSpace,
     type GridColDef,
@@ -11,14 +12,12 @@ import {
     muiGridSortToGql,
     StackLink,
     TableDeleteButton,
-    ToolbarActions,
-    ToolbarItem,
     useBufferedRowCount,
     useDataGridRemote,
     usePersistentColumnState,
 } from "@comet/admin";
 import { Add as AddIcon, Delete as DeleteIcon, Edit } from "@comet/admin-icons";
-import { Button, IconButton, Typography } from "@mui/material";
+import { IconButton, Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { DataGrid, getGridSingleSelectOperators, GridToolbarQuickFilter } from "@mui/x-data-grid";
 import { FormattedMessage, useIntl } from "react-intl";
@@ -37,18 +36,12 @@ interface Props {
 function RedirectsGridToolbar() {
     return (
         <DataGridToolbar>
-            <ToolbarItem>
-                <GridToolbarQuickFilter />
-            </ToolbarItem>
-            <ToolbarItem>
-                <GridFilterButton />
-            </ToolbarItem>
+            <GridToolbarQuickFilter />
+            <GridFilterButton />
             <FillSpace />
-            <ToolbarActions>
-                <Button startIcon={<AddIcon />} component={StackLink} pageName="add" payload="add" variant="contained" color="primary">
-                    <FormattedMessage id="comet.pages.redirects.add" defaultMessage="New redirect" />
-                </Button>
-            </ToolbarActions>
+            <Button startIcon={<AddIcon />} component={StackLink} pageName="add" payload="add">
+                <FormattedMessage id="comet.pages.redirects.add" defaultMessage="New redirect" />
+            </Button>
         </DataGridToolbar>
     );
 }
@@ -92,7 +85,6 @@ export function RedirectsGrid({ linkBlock, scope }: Props): JSX.Element {
             },
             sortable: false,
             flex: 2,
-            filterable: false,
         },
         {
             field: "comment",
@@ -134,6 +126,7 @@ export function RedirectsGrid({ linkBlock, scope }: Props): JSX.Element {
         },
         {
             field: "actions",
+            type: "actions",
             headerName: "",
             renderCell: (params) => (
                 <IconWrapper>