+++ storybook/src/admin/stack/StackTableFormQueryAtStack.stories.tsx
@@ -95,7 +95,7 @@ function ExampleTable(props: IExampleTableProps) {
                             cellProps: { padding: "none" },
 
                             render: (row) => (
-                                <Grid item>
+                                <Grid>
                                     <IconButton
                                         onClick={() => {
                                             stackApi.activatePage("form", String(row.id));