+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @comet/eslint-config
 
+## 8.0.0-beta.3
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.0.0-beta.3
+
+## 8.0.0-beta.2
+
+### Major Changes
+
+- f904b71: Require Node v22
+
+    The minimum required Node version is now v22.0.0.
+    See the migration guide for instructions on how to upgrade your project.
+
+### Patch Changes
+
+- Updated dependencies [f904b71]
+    - @comet/eslint-plugin@8.0.0-beta.2
+
 ## 8.0.0-beta.1
 
 ### Patch Changes
@@ -26,6 +46,18 @@
 ### Minor Changes
 
 - a8edddb: Enable `@typescript-eslint/consistent-type-imports` rule
+
+## 7.17.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@7.17.0
+
+## 7.16.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@7.16.0
 
 ## 7.15.0